… src/Benchmarks/ExperimentalBenchmark/Benchmark.fbs
… src/Benchmarks/ExperimentalBenchmark/ExperimentalBenchmark.csproj
… src/Benchmarks/ExperimentalBenchmark/Program.cs
… src/FlatSharp.Runtime/StringSpanComparer.cs
… src/FlatSharp.Runtime/Vectors/FlatBufferProgressiveVector.cs
… src/FlatSharp.Runtime/Vectors/FlatBufferVectorBase.cs
… src/FlatSharp.Runtime/Vectors/IndexedVector.cs
… src/FlatSharp/Serialization/DeserializeClassDefinition.cs
… src/Tests/Coverage/Coverage.csproj
… src/Tests/FlatSharpTests/OracleTests/OracleDeserializeTests.cs
… src/Tests/FlatSharpTests/SerializationTests/VectorSerializationTests.cs