Optimize memory allocations during binary search:

- No more delegate allocations (structs instead!)
- In non-greedy modes, add a fast-path binary search that avoids the `byte` -> `string` -> `byte` indirection that is necessary if going through the deserialized type.
- Make use of `ArrayPool<byte>.Shared` to avoid allocating temporary buffers

Results:
```
original
|           Method |        Option | Length | TravCount |     Mean |     Error |    StdDev |     Gen 0 |    Gen 1 |    Allocated |
|----------------- |-------------- |------- |---------- |---------:|----------:|----------:|----------:|---------:|-------------:|
|        Serialize |          Lazy |  10000 |         1 | 1.603 ms | 0.0226 ms | 0.0269 ms |         - |        - |         90 B |
| ParseAndTraverse |          Lazy |  10000 |         1 | 8.921 ms | 0.2489 ms | 0.2963 ms | 1328.1250 |        - | 22,442,928 B |
|        Serialize | GreedyMutable |  10000 |         1 | 1.585 ms | 0.0167 ms | 0.0199 ms |         - |        - |         90 B |
| ParseAndTraverse | GreedyMutable |  10000 |         1 | 4.736 ms | 0.0955 ms | 0.1137 ms |  296.8750 | 125.0000 |  5,040,144 B |
```
```
with fixes
|           Method |        Option | Length | TravCount |     Mean |     Error |    StdDev |   Gen 0 |   Gen 1 |   Allocated |
|----------------- |-------------- |------- |---------- |---------:|----------:|----------:|--------:|--------:|------------:|
|        Serialize |          Lazy |  10000 |         1 | 1.459 ms | 0.0154 ms | 0.0040 ms |       - |       - |        90 B |
| ParseAndTraverse |          Lazy |  10000 |         1 | 3.967 ms | 0.1170 ms | 0.0181 ms | 39.0625 |       - |   720,107 B |
|        Serialize | GreedyMutable |  10000 |         1 | 1.377 ms | 0.0014 ms | 0.0004 ms |       - |       - |        90 B |
| ParseAndTraverse | GreedyMutable |  10000 |         1 | 4.849 ms | 0.0348 ms | 0.0054 ms | 78.1250 | 31.2500 | 1,360,147 B |
```